Extract public pad menu helper
