+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{M as ge,a as ht,b as V,A as d,c as b,o as G,n as bt,d as gt}from"./oc-compat-CCOQZroc.chunk.mjs";import{p as q,a as vt}from"./urls-COV-UP-R.chunk.mjs";const L="data-epnc-single-file-share";let ve=0;const Ce=p=>{if(!p)return null;const f='a[href*="/s/'+p+'/download"],a[href*="/index.php/s/'+p+'/download"]',m=document.querySelector(f);return m instanceof HTMLAnchorElement?m:null},Ct=()=>{const p=document.getElementById("initial-state-files_sharing-view");if(!(p instanceof HTMLInputElement))return"";const f=String(p.value||"").trim();if(f==="")return"";try{const m=window.atob(f);if(m.startsWith('"')||m.startsWith("{")||m.startsWith("[")){const v=JSON.parse(m);return typeof v=="string"?v:""}return m}catch{return""}},Ee=p=>{if(!p)return!1;const f=Ce(p);if(!f)return!1;let m;try{m=new URL(f.getAttribute("href")||"",window.location.origin)}catch{return!1}const v=(m.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!v||v[1]!==p?!1:!m.searchParams.has("files")},Et=()=>{const p=document.body;if(!(p instanceof HTMLElement))return{retry:!0};const f=q();if(!f)return p.removeAttribute(L),{retry:!1};const m=Ct();return m==="public-file-share"?(p.setAttribute(L,"1"),{retry:!1}):m==="public-share"?(p.removeAttribute(L),{retry:!1}):Ce(f)?Ee(f)?(p.setAttribute(L,"1"),{retry:!1}):(p.removeAttribute(L),{retry:!1}):(p.removeAttribute(L),{retry:!0})},xt=()=>{const p=++ve,f=m=>{p===ve&&Et().retry&&m<30&&window.setTimeout(()=>f(m+1),180)};f(0)};(function(){const p="etherpad_nextcloud_open",f=new Map;let m=!1,v=null,J=null,Z=0,Q="",S=null,O=null,_=0,H=!1,C=!1,X=!1,Y=0;const B="data-epnc-sidebar-sync-panel",$="data-epnc-sidebar-sync-mount",ee=8e3,xe=6e4,te=15e3,Ae=8e3,ne=d+"_public_pad",ie="icon-filetype-etherpad-nextcloud-pad",re='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',oe=98,Ne=120,Le=500,ae=()=>{const e=bt();return e!==null&&e>=33},A=(e,n)=>{const i=!e||e==="/"?"":String(e),r=String(n||"").replace(/^\/+/,"").replace(/\s+\.pad$/i,".pad");return i===""?"/"+r:i+"/"+r},Se=e=>b("/apps/"+d+"/?file="+encodeURIComponent(e)),N=(e,n)=>{const i=b("/apps/files/files/"+encodeURIComponent(String(e))),r=new URLSearchParams;return r.set("dir",de(n)),r.set("editing","false"),r.set("openfile","true"),i+"?"+r.toString()},k=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),w=()=>(window.location.pathname||"").includes("/apps/files"),se=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=n.pathname||"",r=i.match(/\/apps\/files\/files\/(\d+)\/?$/);if(!r){const a=i.match(/\/f\/(\d+)\/?$/);if(!a){const l=Number(n.searchParams.get("fileid")||"");return Number.isFinite(l)&&l>0?l:null}const s=parseInt(a[1],10);return Number.isFinite(s)&&s>0?s:null}const o=parseInt(r[1],10);return Number.isFinite(o)&&o>0?o:null},I=()=>{const e=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!e)return null;const n=parseInt(e[1],10);return Number.isFinite(n)&&n>0?n:null},R=(e,n)=>{const i=b("/apps/"+d+"/public/"+encodeURIComponent(e));return n?i+"?file="+encodeURIComponent(n):i},ke=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);if(!i)return null;const r=(n.searchParams.get("files")||"").trim();if(!M(r))return null;const o=(n.searchParams.get("path")||"/").trim()||"/";return{token:i[1],path:A(o,r)}},M=e=>typeof e=="string"&&e.toLowerCase().endsWith(".pad"),T=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),Ie=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const n=gt(),i=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of i){const o=Number(r);if(Number.isFinite(o))return(o&n)===n}return null},Me=(...e)=>{for(const n of e){const i=Ie(n);if(i!==null)return i}return!0},Te=(...e)=>Me(...e),le=()=>{it()},ce=()=>({id:ne,displayName:t(d,"Public pad"),...ae()?{iconSvgInline:re}:{iconClass:ie},order:oe,enabled:Te,handler:le}),Fe=()=>({id:ne,displayName:t(d,"Public pad"),templateName:t(d,"Public pad"),iconClass:ie,...ae()?{iconSvgInline:re}:{},fileType:"file",order:oe,actionHandler:le}),U=async e=>{const n=String(e);if(f.has(n))return f.get(n);const i=b("/apps/"+d+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},Pe=async e=>{const n="path:"+String(e);if(f.has(n))return f.get(n);const i=b("/apps/"+d+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve by path failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},qe=e=>{if(!e||typeof e!="string")return"/";const n=e.lastIndexOf("/");return n<=0?"/":e.slice(0,n)||"/"},de=e=>{const n=qe(e);if(n!=="/")return n;const i=j();return i===""?"/":i},j=()=>{const e=(new URLSearchParams(window.location.search||"").get("dir")||"/").trim();return e===""?"/":e},Oe=()=>{const e=document.documentElement,n=document.body,i=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(i))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},ue=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},_e=()=>{const e=I();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),i=Number(n.get("fileid")||"");return Number.isFinite(i)&&i>0?i:null},z=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},He=e=>{if(!e)return null;const n=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const i of n){const r=z(i);if(r!==null)return r}if(typeof e.get=="function")for(const i of["fileid","fileId","id"])try{const r=z(e.get(i));if(r!==null)return r}catch{}return null},Re=e=>{if(!(e instanceof Element))return null;const n=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of n){const o=z(r);if(o!==null)return o}const i=e.matches("a[href]")?e:e.querySelector("a[href]");return i instanceof HTMLAnchorElement?se(i.getAttribute("href")):null},Ue=e=>{if(!e||typeof e!="object")return null;const n=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of n){const o=He(r);if(o!==null)return o}const i=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of i){const o=Re(r);if(o!==null)return o}return null},pe=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const i=document.querySelector(n);if(i instanceof HTMLElement)return i}return null},je=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+$+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute($,"1"));const i=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(i&&i.parentNode instanceof HTMLElement){const r=i.parentNode;return(n.parentNode!==r||n.nextSibling!==i)&&r.insertBefore(n,i),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},fe=e=>{const n=_e();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const i=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of i){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const s=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const l of s){const c=z(l);if(c!==null)return c}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=se(o.getAttribute("href"));if(a!==null)return a}return null},ze=e=>e==="synced"?"etherpad-icon-color":Oe()?"etherpad-icon-white":"etherpad-icon-black",Be=e=>e==="synced"?t(d,"Synchronized"):e==="syncing"?t(d,"Saving..."):e==="error"?t(d,"Sync failed"):e==="unavailable"?t(d,"Sync status unavailable"):e==="loading"?t(d,"Checking sync status..."):t(d,"Sync pending"),x=()=>{const e=document.querySelector("["+B+'="1"]');e&&e.parentNode&&e.parentNode.removeChild(e);const n=document.querySelector("["+$+'="1"]');n&&n.parentNode&&n.parentNode.removeChild(n),F()},$e=(e,n)=>{if(!(e instanceof HTMLElement))return;const i=String(n||"").trim();e.setAttribute("data-public-open-url",i);const r=e.querySelector("[data-epnc-sidebar-open-public-button]");if(!(r instanceof HTMLButtonElement))return;const o=i!=="";r.style.display=o?"inline-flex":"none",r.disabled=!o},E=(e,n,i,r,o=!0)=>{if(!(e instanceof HTMLElement))return;const a=e.querySelector("[data-epnc-sidebar-sync-icon]"),s=e.querySelector("[data-epnc-sidebar-sync-label]"),l=e.querySelector("[data-epnc-sidebar-sync-button]");if(e.setAttribute("data-file-id",String(i)),e.setAttribute("data-sync-state",n),e.setAttribute("data-sync-in-flight",r?"1":"0"),s instanceof HTMLElement&&(s.textContent=Be(n)),a instanceof HTMLElement){const c=ze(n),u=V(d,c);u!==""&&(a.style.backgroundImage=`url(${u})`)}l instanceof HTMLButtonElement&&(l.disabled=r||!o||i<=0,l.setAttribute("aria-busy",r?"true":"false"))},De=(e,n)=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+B+'="1"]');if(!(i instanceof HTMLElement)){i=document.createElement("section"),i.className="epnc-sidebar-sync-panel",i.setAttribute(B,"1"),i.setAttribute("data-file-id",String(e));const r=document.createElement("div");r.className="epnc-sidebar-sync-panel__header";const o=document.createElement("span");o.className="epnc-sidebar-sync-panel__icon",o.setAttribute("data-epnc-sidebar-sync-icon","1"),o.setAttribute("aria-hidden","true");const a=document.createElement("span");a.className="epnc-sidebar-sync-panel__status",a.setAttribute("data-epnc-sidebar-sync-label","1"),r.appendChild(o),r.appendChild(a);const s=document.createElement("button");s.type="button",s.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",s.setAttribute("data-epnc-sidebar-sync-button","1"),s.textContent=t(d,"Pad in Datei speichern"),s.addEventListener("click",async c=>{c.preventDefault(),c.stopPropagation();const u=Number(i.getAttribute("data-file-id")||"");if(!(!Number.isFinite(u)||u<=0)){E(i,"syncing",u,!0,!0);try{await We(u,!0);const h=await W(u),g=D(h&&h.status);E(i,g,u,!1,!0)}catch{E(i,"error",u,!1,!0)}}});const l=document.createElement("button");l.type="button",l.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",l.setAttribute("data-epnc-sidebar-open-public-button","1"),l.textContent=t(d,"Pad in neuem Tab öffnen"),l.style.display="none",l.addEventListener("click",c=>{c.preventDefault(),c.stopPropagation();const u=String(i.getAttribute("data-public-open-url")||"").trim();u!==""&&window.open(u,"_blank","noopener,noreferrer")}),i.appendChild(r),i.appendChild(s),i.appendChild(l)}return i.parentNode!==n&&n.appendChild(i),i},D=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",W=async e=>{const n=b("/apps/"+d+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e))),i=await fetch(n,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}),r=await i.json().catch(()=>({}));if(!i.ok)throw new Error(r&&r.message||"Sync status check failed.");return r},We=async(e,n)=>{let i=b("/apps/"+d+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));i+="?force=1";const r=await fetch(i,{method:"POST",headers:{Accept:"application/json",requesttoken:G()},credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Pad sync failed.");return o},F=()=>{O!==null&&(window.clearTimeout(O),O=null)},Ke=(e,n)=>e==="synced"?Math.min(xe,n+Ae):e==="pending"?ee:te,me=(e,n)=>{if(F(),!(e instanceof HTMLElement)||!Number.isFinite(n)||n<=0)return;let i=ee;const r=a=>{O=window.setTimeout(()=>{o()},Math.max(1e3,a))},o=async()=>{if(!(e instanceof HTMLElement)||!document.body.contains(e)){F();return}const a=pe();if(!(a instanceof HTMLElement)||!ue(a)||!ue(e)){F();return}if(fe(a)!==n){F();return}if(document.visibilityState!=="visible"||e.getAttribute("data-sync-in-flight")==="1"){r(i);return}try{const s=await W(n),l=D(s&&s.status);E(e,l,n,!1,!0),i=Ke(l,i)}catch{E(e,"error",n,!1,!0),i=te}r(i)};r(i)},Ve=()=>{S!==null&&(window.clearTimeout(S),S=null)},Ge=async()=>{const e=++_;if(!w()||q()){x();return}const n=pe();if(!(n instanceof HTMLElement)){x();return}const i=fe(n);if(!i){x();return}const r=je(n);if(!(r instanceof HTMLElement)){x();return}let o;try{o=await U(i)}catch{x();return}if(e!==_)return;if(!o||o.is_pad!==!0){x();return}const a=De(i,r);if(!(a instanceof HTMLElement)){x();return}const s=o&&typeof o.public_open_url=="string"?o.public_open_url.trim():"";$e(a,s),E(a,"loading",i,!1,!0);try{const l=await W(i);if(e!==_)return;const c=D(l&&l.status);E(a,c,i,!1,!0),me(a,i)}catch{if(e!==_)return;E(a,"error",i,!1,!0),me(a,i)}},K=(e=0)=>{Ve(),S=window.setTimeout(()=>{S=null,Ge()},Math.max(0,e))},Je=e=>{try{const n=new URL(e,window.location.origin),i=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(i&&i[1]){const r=i[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},Ze=async(e,n)=>{const i=b("/apps/"+d+"/api/v1/pads/from-url"),r=new URLSearchParams;r.set("file",e),r.set("padUrl",n);const o=await fetch(i,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:G()},credentials:"same-origin",body:r.toString()}),a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Could not import public pad URL.");return a},Qe=async e=>{const n=b("/apps/"+d+"/api/v1/pads"),i=new URLSearchParams;i.set("file",e),i.set("accessMode","public");const r=await fetch(n,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:G()},credentials:"same-origin",body:i.toString()}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Could not create public pad.");return o},ye=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(d,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",i.appendChild(r),i.appendChild(o),n.appendChild(i),document.body.appendChild(n),{overlay:n,dialog:i,closeButton:r}},Xe=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=ye(t(d,"Internal public pad")),o=document.createElement("label");o.textContent=t(d,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(d,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("p");s.style.color="var(--color-error, #c62828)",s.style.margin="0 0 12px 0",s.style.minHeight="20px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Create");const c=u=>{n.remove(),e(u)};r.addEventListener("click",()=>c(null)),n.addEventListener("click",u=>{u.target===n&&c(null)}),a.addEventListener("keydown",u=>{u.key==="Enter"&&(u.preventDefault(),l.click())}),l.addEventListener("click",()=>{const u=a.value.trim();if(u===""){s.textContent=t(d,"File name is required.");return}c(u)}),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),a.focus(),a.select()}),Ye=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=ye(t(d,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(d,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const s=document.createElement("label");s.textContent=t(d,"File name"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="text",l.value="Imported pad.pad",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const c=document.createElement("p");c.style.color="var(--color-error, #c62828)",c.style.margin="0 0 12px 0",c.style.minHeight="20px";const u=document.createElement("button");u.type="button",u.className="primary",u.textContent=t(d,"Create");const h=y=>{n.remove(),e(y)};r.addEventListener("click",()=>h(null)),n.addEventListener("click",y=>{y.target===n&&h(null)}),a.addEventListener("blur",()=>{const y=a.value.trim();y.startsWith("http")&&(l.value=Je(y))});const g=()=>{const y=a.value.trim(),be=l.value.trim();if(y===""){c.textContent=t(d,"External pad URL is required.");return}if(be===""){c.textContent=t(d,"File name is required.");return}h({padUrl:y,name:be})};a.addEventListener("keydown",y=>{y.key==="Enter"&&(y.preventDefault(),g())}),l.addEventListener("keydown",y=>{y.key==="Enter"&&(y.preventDefault(),g())}),u.addEventListener("click",g),i.appendChild(o),i.appendChild(a),i.appendChild(s),i.appendChild(l),i.appendChild(c),i.appendChild(u),a.focus(),a.select()}),et=async()=>{const e=await Xe();if(!e)return;const n=e.trim(),i=n.toLowerCase().endsWith(".pad")?n:n+".pad",r=A(j(),i);try{const o=await Qe(r),a=o&&typeof o.file=="string"?o.file:r,s=o&&Number.isFinite(Number(o.file_id))?Number(o.file_id):null;if(w()&&s!==null){window.location.assign(N(s,a));return}await P({path:a,fileId:s})}catch(o){const a=o instanceof Error?o.message:t(d,"Could not create public pad.");window.alert(a)}},tt=async()=>{const e=await Ye();if(!e)return;const n=e.padUrl.trim(),i=e.name.trim(),r=j(),o=i.toLowerCase().endsWith(".pad")?i:i+".pad",a=A(r,o);try{const s=await Ze(a,n),l=s&&typeof s.file=="string"?s.file:a,c=s&&Number.isFinite(Number(s.file_id))?Number(s.file_id):null;if(w()&&c!==null){window.location.assign(N(c,l));return}await P({path:l,fileId:c})}catch(s){const l=s instanceof Error?s.message:t(d,"Could not import public pad URL.");window.alert(l)}},nt=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(d,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(d,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(d,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const s=document.createElement("div");s.style.display="grid",s.style.gap="8px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(d,"Internal public pad");const c=document.createElement("button");c.type="button",c.textContent=t(d,"External public pad (URL)");const u=V(d,"etherpad-icon-color");u!==""&&(c.style.backgroundImage=`url(${u})`),c.style.backgroundRepeat="no-repeat",c.style.backgroundPosition="12px center",c.style.backgroundSize="16px 16px",c.style.paddingLeft="34px";const h=g=>{n.remove(),e(g)};o.addEventListener("click",()=>h(null)),l.addEventListener("click",()=>h("internal")),c.addEventListener("click",()=>h("external")),n.addEventListener("click",g=>{g.target===n&&h(null)}),s.appendChild(l),s.appendChild(c),i.appendChild(o),i.appendChild(r),i.appendChild(a),i.appendChild(s),n.appendChild(i),document.body.appendChild(n)}),it=async()=>{const e=await nt();if(e==="internal"){await et();return}e==="external"&&await tt()},rt=(e,n)=>{const i=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!i||typeof i.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(i.goToRoute(null,{view:"files",fileid:String(e)},{dir:de(n),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=N(e,n);window.location.assign(r)},180),!0)},P=async e=>{const n=String(e.fileId??"")+"|"+String(e.path??""),i=Date.now();if(J===n&&i-Z<800)return;J=n,Z=i;const r=q(),o=r!==null&&r!=="",a=()=>{if(o){window.location.assign(R(r,e.path||""));return}if(w()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(N(Number(e.fileId),u));return}const c=w()?I():null;if((e.fileId===null||e.fileId===void 0)&&c){const u=e.path||"/";window.location.assign(N(c,u));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const u=e.path||"/";window.location.assign(N(Number(e.fileId),u));return}e.path&&window.location.assign(Se(e.path))};if(!k()){a();return}let s=e.path||"",l=e.fileId??null;if(!s&&e.fileId!==null&&e.fileId!==void 0)try{const c=await U(e.fileId);s=c&&typeof c.path=="string"?c.path:"",l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{s=""}if(!s){a();return}if((!l||!Number.isFinite(Number(l)))&&s&&!o)try{const c=await Pe(s);l=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):l}catch{}if((!l||!Number.isFinite(Number(l)))&&w()){const c=I();c&&(l=c)}if(w()){if(l&&Number.isFinite(Number(l))&&rt(Number(l),s))return;a();return}try{window.OCA.Viewer.open({path:s})}catch{a()}},ot=(e,n)=>{if(!(e instanceof HTMLAnchorElement)||!n)return null;const i=e.getAttribute("href")||"",r=ke(i);if(r&&r.token===n&&M(r.path))return r.path;const o=vt(i);return M(o)?o:null},at=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const e=async n=>{if(!n||n.defaultPrevented)return;const i=q();if(!i||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const r=n.target.closest("a[href]");if(!(r instanceof HTMLAnchorElement))return;const o=ot(r,i);if(o){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),k()){await P({path:o,fileId:null});return}window.location.assign(R(i,o))}};document.addEventListener("click",e,!0)},st=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ge,p,ht(),V(d,"etherpad-icon-color"),(e,n)=>{const i=n&&n.dir||j(),r=A(i,e),o=Ue(n);P({path:r,fileId:o})},t(d,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ge,p))},lt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const i of e)if(!(!i||typeof i!="object")){if(typeof i.addNewFileMenuEntry=="function")return{register:i.addNewFileMenuEntry.bind(i),unregister:typeof i.removeNewFileMenuEntry=="function"?i.removeNewFileMenuEntry.bind(i):null};if(typeof i.getNewFileMenu=="function"){let r=null;try{r=i.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const n=window._nc_newfilemenu;return n&&typeof n=="object"&&typeof n.registerEntry=="function"?{register:n.registerEntry.bind(n),unregister:typeof n.unregisterEntry=="function"?n.unregisterEntry.bind(n):null}:null},ct=async()=>{const e=lt();if(!e)return!1;const n=ce();try{return e.register(n),H=!0,!0}catch(i){return T(i)?(H=!0,!0):!1}},dt=()=>{if(C)return!0;const e=Fe(),n=()=>{const i=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const r of i)if(!(!r||typeof r!="object")){if(typeof r.addMenuEntry=="function")try{return r.addMenuEntry(e),!0}catch(o){if(T(o))return!0}if(typeof r.registerEntry=="function")try{return r.registerEntry(ce()),!0}catch(o){if(T(o))return!0}}return!1};try{return n()?(C=!0,!0):(!X&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(i){if(!(!i||typeof i.addMenuEntry!="function"))try{i.addMenuEntry(e),C=!0}catch(r){T(r)&&(C=!0)}}}),X=!0),C)}catch(i){return T(i)?(C=!0,!0):!1}},ut=()=>{if(!w()||H||C)return;const e=++Y,n=async i=>{if(!(e!==Y||H||C||await ct()||dt())&&i<Ne){window.setTimeout(()=>{n(i+1)},Le);return}};n(0)},pt=()=>{if(v!==null||!w())return;const e=new MutationObserver(n=>{for(const i of n)for(const r of i.addedNodes)if(r instanceof Element&&(r.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||r.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){K(60);return}});e.observe(document.body,{childList:!0,subtree:!0}),v=e},ft=async()=>{if(k())return;const e=I();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("epNoRedirect")==="1"||n.get("details")==="1"||n.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+d+"/"))try{const i=await U(e);i&&i.is_pad&&i.viewer_url&&window.location.assign(i.viewer_url)}catch{}},mt=async()=>{if(k()||!w())return;const e=I();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("openfile")==="true"||n.get("opendetails")==="true"||n.get("details")==="1"))try{const i=await U(e);if(!i||!i.is_pad)return;const r=n.get("dir")||"/",o=b("/apps/files/files")+"?dir="+encodeURIComponent(r);window.location.href!==o&&window.location.assign(o)}catch{}},yt=()=>{const e=q();if(!e)return;const n=new URLSearchParams(window.location.search||""),i=n.get("files")||"";if(k()){if(!M(i))return;const a=n.get("path")||"/",s=A(a,i);P({path:s,fileId:null});return}if(!M(i)){if((window.location.pathname||"").includes("/apps/"+d+"/public/"))return;Ee(e)&&window.location.assign(R(e,""));return}const r=n.get("path")||"/",o=A(r,i);window.location.assign(R(e,o))},we=()=>{const e=(window.location.pathname||"")+"?"+(window.location.search||"");e!==Q&&(Q=e,w()&&ut(),xt(),mt(),ft(),yt(),K(0))},wt=()=>{const e="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[e]===!0)return;window.OCA&&(window.OCA[e]=!0);const n=()=>window.setTimeout(we,0);window.addEventListener("popstate",n),window.addEventListener("hashchange",n);const i=r=>{const o=window.history&&window.history[r];typeof o=="function"&&(window.history[r]=function(...a){const s=o.apply(this,a);return n(),s})};i("pushState"),i("replaceState")},he=()=>{m||(m=!0,wt(),we(),st(),at(),pt(),K(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",he,{once:!0}):he()})();
+import{A as u,n as it,a as rt,M as ue,b as ot,c as X,d as E,o as Y}from"./oc-compat-DyRpMLpS.chunk.mjs";import{p as B,a as at}from"./urls-COV-UP-R.chunk.mjs";const pe=u+"_public_pad",fe="icon-filetype-etherpad-nextcloud-pad",me='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',ye=98,st=120,lt=500,we=()=>{const d=it();return d!==null&&d>=33},z=d=>(d&&d.message?String(d.message):"").toLowerCase().includes("duplicate"),ct=d=>{if(!d||typeof d!="object")return null;if(d.hasCreatePermission===!1)return!1;if(d.hasCreatePermission===!0)return!0;if(typeof d.canCreate=="boolean")return d.canCreate;const f=rt(),p=[d.permissions,d.permission,d.attributes&&d.attributes.permissions,typeof d.get=="function"?d.get("permissions"):null];for(const y of p){const N=Number(y);if(Number.isFinite(N))return(N&f)===f}return null},dt=(...d)=>{for(const f of d){const p=ct(f);if(p!==null)return p}return!0},ut=()=>{const d=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const p of d)if(!(!p||typeof p!="object")){if(typeof p.addNewFileMenuEntry=="function")return{register:p.addNewFileMenuEntry.bind(p),unregister:typeof p.removeNewFileMenuEntry=="function"?p.removeNewFileMenuEntry.bind(p):null};if(typeof p.getNewFileMenu=="function"){let y=null;try{y=p.getNewFileMenu()}catch{y=null}if(y&&typeof y.registerEntry=="function")return{register:y.registerEntry.bind(y),unregister:typeof y.unregisterEntry=="function"?y.unregisterEntry.bind(y):null}}}const f=window._nc_newfilemenu;return f&&typeof f=="object"&&typeof f.registerEntry=="function"?{register:f.registerEntry.bind(f),unregister:typeof f.unregisterEntry=="function"?f.unregisterEntry.bind(f):null}:null},pt=({isFilesAppRoute:d,onCreatePublicPad:f})=>{let p=!1,y=!1,N=!1,q=0;const O=()=>{f()},k=(...b)=>dt(...b),I=()=>({id:pe,displayName:t(u,"Public pad"),...we()?{iconSvgInline:me}:{iconClass:fe},order:ye,enabled:k,handler:O}),P=()=>({id:pe,displayName:t(u,"Public pad"),templateName:t(u,"Public pad"),iconClass:fe,...we()?{iconSvgInline:me}:{},fileType:"file",order:ye,actionHandler:O}),_=async()=>{const b=ut();if(!b)return!1;const L=I();try{return b.register(L),p=!0,!0}catch(w){return z(w)?(p=!0,!0):!1}},R=()=>{if(y)return!0;const b=P(),L=()=>{const w=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const v of w)if(!(!v||typeof v!="object")){if(typeof v.addMenuEntry=="function")try{return v.addMenuEntry(b),!0}catch(C){if(z(C))return!0}if(typeof v.registerEntry=="function")try{return v.registerEntry(I()),!0}catch(C){if(z(C))return!0}}return!1};try{return L()?(y=!0,!0):(!N&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(w){if(!(!w||typeof w.addMenuEntry!="function"))try{w.addMenuEntry(b),y=!0}catch(v){z(v)&&(y=!0)}}}),N=!0),y)}catch(w){return z(w)?(y=!0,!0):!1}};return()=>{if(!d()||p||y)return;const b=++q,L=async w=>{b!==q||p||y||await _()||R()||w<st&&window.setTimeout(()=>{L(w+1)},lt)};L(0)}},F="data-epnc-single-file-share";let he=0;const be=d=>{if(!d)return null;const f='a[href*="/s/'+d+'/download"],a[href*="/index.php/s/'+d+'/download"]',p=document.querySelector(f);return p instanceof HTMLAnchorElement?p:null},ft=()=>{const d=document.getElementById("initial-state-files_sharing-view");if(!(d instanceof HTMLInputElement))return"";const f=String(d.value||"").trim();if(f==="")return"";try{const p=window.atob(f);if(p.startsWith('"')||p.startsWith("{")||p.startsWith("[")){const y=JSON.parse(p);return typeof y=="string"?y:""}return p}catch{return""}},ge=d=>{if(!d)return!1;const f=be(d);if(!f)return!1;let p;try{p=new URL(f.getAttribute("href")||"",window.location.origin)}catch{return!1}const y=(p.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!y||y[1]!==d?!1:!p.searchParams.has("files")},mt=()=>{const d=document.body;if(!(d instanceof HTMLElement))return{retry:!0};const f=B();if(!f)return d.removeAttribute(F),{retry:!1};const p=ft();return p==="public-file-share"?(d.setAttribute(F,"1"),{retry:!1}):p==="public-share"?(d.removeAttribute(F),{retry:!1}):be(f)?ge(f)?(d.setAttribute(F,"1"),{retry:!1}):(d.removeAttribute(F),{retry:!1}):(d.removeAttribute(F),{retry:!0})},yt=()=>{const d=++he,f=p=>{d===he&&mt().retry&&p<30&&window.setTimeout(()=>f(p+1),180)};f(0)};(function(){const d="etherpad_nextcloud_open",f=new Map;let p=!1,y=null,N=null,q=0,O="",k=null,I=null,P=0;const _="data-epnc-sidebar-sync-panel",R="data-epnc-sidebar-sync-mount",b=8e3,L=6e4,w=15e3,v=8e3,C=(e,n)=>{const i=!e||e==="/"?"":String(e),r=String(n||"").replace(/^\/+/,"").replace(/\s+\.pad$/i,".pad");return i===""?"/"+r:i+"/"+r},ve=e=>E("/apps/"+u+"/?file="+encodeURIComponent(e)),$=(e,n)=>{const i=E("/apps/files/files/"+encodeURIComponent(String(e))),r=new URLSearchParams;return r.set("dir",te(n)),r.set("editing","false"),r.set("openfile","true"),i+"?"+r.toString()},T=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),x=()=>(window.location.pathname||"").includes("/apps/files"),ee=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=n.pathname||"",r=i.match(/\/apps\/files\/files\/(\d+)\/?$/);if(!r){const a=i.match(/\/f\/(\d+)\/?$/);if(!a){const s=Number(n.searchParams.get("fileid")||"");return Number.isFinite(s)&&s>0?s:null}const l=parseInt(a[1],10);return Number.isFinite(l)&&l>0?l:null}const o=parseInt(r[1],10);return Number.isFinite(o)&&o>0?o:null},H=()=>{const e=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!e)return null;const n=parseInt(e[1],10);return Number.isFinite(n)&&n>0?n:null},D=(e,n)=>{const i=E("/apps/"+u+"/public/"+encodeURIComponent(e));return n?i+"?file="+encodeURIComponent(n):i},Ce=e=>{if(!e||typeof e!="string")return null;let n;try{n=new URL(e,window.location.origin)}catch{return null}const i=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);if(!i)return null;const r=(n.searchParams.get("files")||"").trim();if(!U(r))return null;const o=(n.searchParams.get("path")||"/").trim()||"/";return{token:i[1],path:C(o,r)}},U=e=>typeof e=="string"&&e.toLowerCase().endsWith(".pad"),W=async e=>{const n=String(e);if(f.has(n))return f.get(n);const i=E("/apps/"+u+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},Ee=async e=>{const n="path:"+String(e);if(f.has(n))return f.get(n);const i=E("/apps/"+u+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=fetch(i,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}).then(async o=>{const a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Pad resolve by path failed.");return a}).catch(o=>{throw f.delete(n),o});return f.set(n,r),r},xe=e=>{if(!e||typeof e!="string")return"/";const n=e.lastIndexOf("/");return n<=0?"/":e.slice(0,n)||"/"},te=e=>{const n=xe(e);if(n!=="/")return n;const i=V();return i===""?"/":i},V=()=>{const e=(new URLSearchParams(window.location.search||"").get("dir")||"/").trim();return e===""?"/":e},Ae=()=>{const e=document.documentElement,n=document.body,i=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(i))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},ne=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},Ne=()=>{const e=H();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),i=Number(n.get("fileid")||"");return Number.isFinite(i)&&i>0?i:null},K=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},Le=e=>{if(!e)return null;const n=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const i of n){const r=K(i);if(r!==null)return r}if(typeof e.get=="function")for(const i of["fileid","fileId","id"])try{const r=K(e.get(i));if(r!==null)return r}catch{}return null},Se=e=>{if(!(e instanceof Element))return null;const n=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of n){const o=K(r);if(o!==null)return o}const i=e.matches("a[href]")?e:e.querySelector("a[href]");return i instanceof HTMLAnchorElement?ee(i.getAttribute("href")):null},ke=e=>{if(!e||typeof e!="object")return null;const n=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of n){const o=Le(r);if(o!==null)return o}const i=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of i){const o=Se(r);if(o!==null)return o}return null},ie=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const i=document.querySelector(n);if(i instanceof HTMLElement)return i}return null},Ie=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+R+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute(R,"1"));const i=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(i&&i.parentNode instanceof HTMLElement){const r=i.parentNode;return(n.parentNode!==r||n.nextSibling!==i)&&r.insertBefore(n,i),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},re=e=>{const n=Ne();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const i=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of i){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const l=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const s of l){const c=K(s);if(c!==null)return c}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=ee(o.getAttribute("href"));if(a!==null)return a}return null},Me=e=>e==="synced"?"etherpad-icon-color":Ae()?"etherpad-icon-white":"etherpad-icon-black",Pe=e=>e==="synced"?t(u,"Synchronized"):e==="syncing"?t(u,"Saving..."):e==="error"?t(u,"Sync failed"):e==="unavailable"?t(u,"Sync status unavailable"):e==="loading"?t(u,"Checking sync status..."):t(u,"Sync pending"),M=()=>{const e=document.querySelector("["+_+'="1"]');e&&e.parentNode&&e.parentNode.removeChild(e);const n=document.querySelector("["+R+'="1"]');n&&n.parentNode&&n.parentNode.removeChild(n),j()},Te=(e,n)=>{if(!(e instanceof HTMLElement))return;const i=String(n||"").trim();e.setAttribute("data-public-open-url",i);const r=e.querySelector("[data-epnc-sidebar-open-public-button]");if(!(r instanceof HTMLButtonElement))return;const o=i!=="";r.style.display=o?"inline-flex":"none",r.disabled=!o},S=(e,n,i,r,o=!0)=>{if(!(e instanceof HTMLElement))return;const a=e.querySelector("[data-epnc-sidebar-sync-icon]"),l=e.querySelector("[data-epnc-sidebar-sync-label]"),s=e.querySelector("[data-epnc-sidebar-sync-button]");if(e.setAttribute("data-file-id",String(i)),e.setAttribute("data-sync-state",n),e.setAttribute("data-sync-in-flight",r?"1":"0"),l instanceof HTMLElement&&(l.textContent=Pe(n)),a instanceof HTMLElement){const c=Me(n),m=X(u,c);m!==""&&(a.style.backgroundImage=`url(${m})`)}s instanceof HTMLButtonElement&&(s.disabled=r||!o||i<=0,s.setAttribute("aria-busy",r?"true":"false"))},Fe=(e,n)=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+_+'="1"]');if(!(i instanceof HTMLElement)){i=document.createElement("section"),i.className="epnc-sidebar-sync-panel",i.setAttribute(_,"1"),i.setAttribute("data-file-id",String(e));const r=document.createElement("div");r.className="epnc-sidebar-sync-panel__header";const o=document.createElement("span");o.className="epnc-sidebar-sync-panel__icon",o.setAttribute("data-epnc-sidebar-sync-icon","1"),o.setAttribute("aria-hidden","true");const a=document.createElement("span");a.className="epnc-sidebar-sync-panel__status",a.setAttribute("data-epnc-sidebar-sync-label","1"),r.appendChild(o),r.appendChild(a);const l=document.createElement("button");l.type="button",l.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",l.setAttribute("data-epnc-sidebar-sync-button","1"),l.textContent=t(u,"Pad in Datei speichern"),l.addEventListener("click",async c=>{c.preventDefault(),c.stopPropagation();const m=Number(i.getAttribute("data-file-id")||"");if(!(!Number.isFinite(m)||m<=0)){S(i,"syncing",m,!0,!0);try{await qe(m,!0);const g=await Z(m),A=J(g&&g.status);S(i,A,m,!1,!0)}catch{S(i,"error",m,!1,!0)}}});const s=document.createElement("button");s.type="button",s.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",s.setAttribute("data-epnc-sidebar-open-public-button","1"),s.textContent=t(u,"Pad in neuem Tab öffnen"),s.style.display="none",s.addEventListener("click",c=>{c.preventDefault(),c.stopPropagation();const m=String(i.getAttribute("data-public-open-url")||"").trim();m!==""&&window.open(m,"_blank","noopener,noreferrer")}),i.appendChild(r),i.appendChild(l),i.appendChild(s)}return i.parentNode!==n&&n.appendChild(i),i},J=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",Z=async e=>{const n=E("/apps/"+u+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e))),i=await fetch(n,{method:"GET",headers:{Accept:"application/json"},credentials:"same-origin"}),r=await i.json().catch(()=>({}));if(!i.ok)throw new Error(r&&r.message||"Sync status check failed.");return r},qe=async(e,n)=>{let i=E("/apps/"+u+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));i+="?force=1";const r=await fetch(i,{method:"POST",headers:{Accept:"application/json",requesttoken:Y()},credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Pad sync failed.");return o},j=()=>{I!==null&&(window.clearTimeout(I),I=null)},Oe=(e,n)=>e==="synced"?Math.min(L,n+v):e==="pending"?b:w,oe=(e,n)=>{if(j(),!(e instanceof HTMLElement)||!Number.isFinite(n)||n<=0)return;let i=b;const r=a=>{I=window.setTimeout(()=>{o()},Math.max(1e3,a))},o=async()=>{if(!(e instanceof HTMLElement)||!document.body.contains(e)){j();return}const a=ie();if(!(a instanceof HTMLElement)||!ne(a)||!ne(e)){j();return}if(re(a)!==n){j();return}if(document.visibilityState!=="visible"||e.getAttribute("data-sync-in-flight")==="1"){r(i);return}try{const l=await Z(n),s=J(l&&l.status);S(e,s,n,!1,!0),i=Oe(s,i)}catch{S(e,"error",n,!1,!0),i=w}r(i)};r(i)},_e=()=>{k!==null&&(window.clearTimeout(k),k=null)},Re=async()=>{const e=++P;if(!x()||B()){M();return}const n=ie();if(!(n instanceof HTMLElement)){M();return}const i=re(n);if(!i){M();return}const r=Ie(n);if(!(r instanceof HTMLElement)){M();return}let o;try{o=await W(i)}catch{M();return}if(e!==P)return;if(!o||o.is_pad!==!0){M();return}const a=Fe(i,r);if(!(a instanceof HTMLElement)){M();return}const l=o&&typeof o.public_open_url=="string"?o.public_open_url.trim():"";Te(a,l),S(a,"loading",i,!1,!0);try{const s=await Z(i);if(e!==P)return;const c=J(s&&s.status);S(a,c,i,!1,!0),oe(a,i)}catch{if(e!==P)return;S(a,"error",i,!1,!0),oe(a,i)}},Q=(e=0)=>{_e(),k=window.setTimeout(()=>{k=null,Re()},Math.max(0,e))},He=e=>{try{const n=new URL(e,window.location.origin),i=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(i&&i[1]){const r=i[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},Ue=async(e,n)=>{const i=E("/apps/"+u+"/api/v1/pads/from-url"),r=new URLSearchParams;r.set("file",e),r.set("padUrl",n);const o=await fetch(i,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:Y()},credentials:"same-origin",body:r.toString()}),a=await o.json().catch(()=>({}));if(!o.ok)throw new Error(a&&a.message||"Could not import public pad URL.");return a},je=async e=>{const n=E("/apps/"+u+"/api/v1/pads"),i=new URLSearchParams;i.set("file",e),i.set("accessMode","public");const r=await fetch(n,{method:"POST",headers:{Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:Y()},credentials:"same-origin",body:i.toString()}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||"Could not create public pad.");return o},ae=async e=>{if(T()&&e.path)try{window.OCA.Viewer.open({path:e.path});return}catch{}await G(e)},se=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(u,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",i.appendChild(r),i.appendChild(o),n.appendChild(i),document.body.appendChild(n),{overlay:n,dialog:i,closeButton:r}},ze=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=se(t(u,"Internal public pad")),o=document.createElement("label");o.textContent=t(u,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(u,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const l=document.createElement("p");l.style.color="var(--color-error, #c62828)",l.style.margin="0 0 12px 0",l.style.minHeight="20px";const s=document.createElement("button");s.type="button",s.className="primary",s.textContent=t(u,"Create");const c=m=>{n.remove(),e(m)};r.addEventListener("click",()=>c(null)),n.addEventListener("click",m=>{m.target===n&&c(null)}),a.addEventListener("keydown",m=>{m.key==="Enter"&&(m.preventDefault(),s.click())}),s.addEventListener("click",()=>{const m=a.value.trim();if(m===""){l.textContent=t(u,"File name is required.");return}c(m)}),i.appendChild(o),i.appendChild(a),i.appendChild(l),i.appendChild(s),a.focus(),a.select()}),Be=()=>new Promise(e=>{const{overlay:n,dialog:i,closeButton:r}=se(t(u,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(u,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const l=document.createElement("label");l.textContent=t(u,"File name"),l.style.display="block",l.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value="Imported pad.pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const c=document.createElement("p");c.style.color="var(--color-error, #c62828)",c.style.margin="0 0 12px 0",c.style.minHeight="20px";const m=document.createElement("button");m.type="button",m.className="primary",m.textContent=t(u,"Create");const g=h=>{n.remove(),e(h)};r.addEventListener("click",()=>g(null)),n.addEventListener("click",h=>{h.target===n&&g(null)}),a.addEventListener("blur",()=>{const h=a.value.trim();h.startsWith("http")&&(s.value=He(h))});const A=()=>{const h=a.value.trim(),de=s.value.trim();if(h===""){c.textContent=t(u,"External pad URL is required.");return}if(de===""){c.textContent=t(u,"File name is required.");return}g({padUrl:h,name:de})};a.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),A())}),s.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),A())}),m.addEventListener("click",A),i.appendChild(o),i.appendChild(a),i.appendChild(l),i.appendChild(s),i.appendChild(c),i.appendChild(m),a.focus(),a.select()}),$e=async()=>{const e=await ze();if(!e)return;const n=e.trim(),i=n.toLowerCase().endsWith(".pad")?n:n+".pad",r=C(V(),i);try{const o=await je(r),a=o&&typeof o.file=="string"?o.file:r,l=o&&Number.isFinite(Number(o.file_id))?Number(o.file_id):null;await ae({path:a,fileId:l})}catch(o){const a=o instanceof Error?o.message:t(u,"Could not create public pad.");window.alert(a)}},De=async()=>{const e=await Be();if(!e)return;const n=e.padUrl.trim(),i=e.name.trim(),r=V(),o=i.toLowerCase().endsWith(".pad")?i:i+".pad",a=C(r,o);try{const l=await Ue(a,n),s=l&&typeof l.file=="string"?l.file:a,c=l&&Number.isFinite(Number(l.file_id))?Number(l.file_id):null;await ae({path:s,fileId:c})}catch(l){const s=l instanceof Error?l.message:t(u,"Could not import public pad URL.");window.alert(s)}},We=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(u,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(u,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(u,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const l=document.createElement("div");l.style.display="grid",l.style.gap="8px";const s=document.createElement("button");s.type="button",s.className="primary",s.textContent=t(u,"Internal public pad");const c=document.createElement("button");c.type="button",c.textContent=t(u,"External public pad (URL)");const m=X(u,"etherpad-icon-color");m!==""&&(c.style.backgroundImage=`url(${m})`),c.style.backgroundRepeat="no-repeat",c.style.backgroundPosition="12px center",c.style.backgroundSize="16px 16px",c.style.paddingLeft="34px";const g=A=>{n.remove(),e(A)};o.addEventListener("click",()=>g(null)),s.addEventListener("click",()=>g("internal")),c.addEventListener("click",()=>g("external")),n.addEventListener("click",A=>{A.target===n&&g(null)}),l.appendChild(s),l.appendChild(c),i.appendChild(o),i.appendChild(r),i.appendChild(a),i.appendChild(l),n.appendChild(i),document.body.appendChild(n)}),Ve=async()=>{const e=await We();if(e==="internal"){await $e();return}e==="external"&&await De()},Ke=(e,n)=>{const i=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!i||typeof i.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(i.goToRoute(null,{view:"files",fileid:String(e)},{dir:te(n),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=$(e,n);window.location.assign(r)},180),!0)},G=async e=>{const n=String(e.fileId??"")+"|"+String(e.path??""),i=Date.now();if(N===n&&i-q<800)return;N=n,q=i;const r=B(),o=r!==null&&r!=="",a=()=>{if(o){window.location.assign(D(r,e.path||""));return}if(x()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const m=e.path||"/";window.location.assign($(Number(e.fileId),m));return}const c=x()?H():null;if((e.fileId===null||e.fileId===void 0)&&c){const m=e.path||"/";window.location.assign($(c,m));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const m=e.path||"/";window.location.assign($(Number(e.fileId),m));return}e.path&&window.location.assign(ve(e.path))};if(!T()){a();return}let l=e.path||"",s=e.fileId??null;if(!l&&e.fileId!==null&&e.fileId!==void 0)try{const c=await W(e.fileId);l=c&&typeof c.path=="string"?c.path:"",s=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):s}catch{l=""}if(!l){a();return}if((!s||!Number.isFinite(Number(s)))&&l&&!o)try{const c=await Ee(l);s=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):s}catch{}if((!s||!Number.isFinite(Number(s)))&&x()){const c=H();c&&(s=c)}if(x()){if(s&&Number.isFinite(Number(s))&&Ke(Number(s),l))return;a();return}try{window.OCA.Viewer.open({path:l})}catch{a()}},Ge=(e,n)=>{if(!(e instanceof HTMLAnchorElement)||!n)return null;const i=e.getAttribute("href")||"",r=Ce(i);if(r&&r.token===n&&U(r.path))return r.path;const o=at(i);return U(o)?o:null},Je=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const e=async n=>{if(!n||n.defaultPrevented)return;const i=B();if(!i||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const r=n.target.closest("a[href]");if(!(r instanceof HTMLAnchorElement))return;const o=Ge(r,i);if(o){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),T()){await G({path:o,fileId:null});return}window.location.assign(D(i,o))}};document.addEventListener("click",e,!0)},Ze=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ue,d,ot(),X(u,"etherpad-icon-color"),(e,n)=>{const i=n&&n.dir||V(),r=C(i,e),o=ke(n);G({path:r,fileId:o})},t(u,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ue,d))},Qe=pt({isFilesAppRoute:x,onCreatePublicPad:Ve}),Xe=()=>{if(y!==null||!x())return;const e=new MutationObserver(n=>{for(const i of n)for(const r of i.addedNodes)if(r instanceof Element&&(r.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||r.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){Q(60);return}});e.observe(document.body,{childList:!0,subtree:!0}),y=e},Ye=async()=>{if(T())return;const e=H();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("epNoRedirect")==="1"||n.get("details")==="1"||n.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+u+"/"))try{const i=await W(e);i&&i.is_pad&&i.viewer_url&&window.location.assign(i.viewer_url)}catch{}},et=async()=>{if(T()||!x())return;const e=H();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("openfile")==="true"||n.get("opendetails")==="true"||n.get("details")==="1"))try{const i=await W(e);if(!i||!i.is_pad)return;const r=n.get("dir")||"/",o=E("/apps/files/files")+"?dir="+encodeURIComponent(r);window.location.href!==o&&window.location.assign(o)}catch{}},tt=()=>{const e=B();if(!e)return;const n=new URLSearchParams(window.location.search||""),i=n.get("files")||"";if(T()){if(!U(i))return;const a=n.get("path")||"/",l=C(a,i);G({path:l,fileId:null});return}if(!U(i)){if((window.location.pathname||"").includes("/apps/"+u+"/public/"))return;ge(e)&&window.location.assign(D(e,""));return}const r=n.get("path")||"/",o=C(r,i);window.location.assign(D(e,o))},le=()=>{const e=(window.location.pathname||"")+"?"+(window.location.search||"");e!==O&&(O=e,x()&&Qe(),yt(),et(),Ye(),tt(),Q(0))},nt=()=>{const e="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[e]===!0)return;window.OCA&&(window.OCA[e]=!0);const n=()=>window.setTimeout(le,0);window.addEventListener("popstate",n),window.addEventListener("hashchange",n);const i=r=>{const o=window.history&&window.history[r];typeof o=="function"&&(window.history[r]=function(...a){const l=o.apply(this,a);return n(),l})};i("pushState"),i("replaceState")},ce=()=>{p||(p=!0,nt(),le(),Ze(),Je(),Xe(),Q(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",ce,{once:!0}):ce()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
