+++ js/oc-compat-DyRpMLpS.chunk.mjs
@@ -1,0 +1,2 @@
+const w="etherpad_nextcloud",d="application/x-etherpad-nextcloud",r="etherpad_nextcloud",c=n=>window.OC&&typeof window.OC.generateUrl=="function"?window.OC.generateUrl(n):"/index.php"+n,a=(n,o)=>window.OC&&typeof window.OC.imagePath=="function"?window.OC.imagePath(n,o):"",s=(n="")=>{const o=String(n||"").trim();return o!==""?o:String(window.OC&&window.OC.requestToken||"")},u=()=>{const n=window.OC&&window.OC.PERMISSION_READ,o=Number(n);return Number.isFinite(o)&&o>0?o:1},O=()=>{const n=window.OC&&window.OC.PERMISSION_CREATE,o=Number(n);return Number.isFinite(o)&&o>0?o:4},p=()=>{const n=[window.OC&&window.OC.config&&window.OC.config.version,window.OC&&window.OC.config&&window.OC.config.versionstring,window.oc_appconfig&&window.oc_appconfig.core&&window.oc_appconfig.core.version,document.documentElement&&document.documentElement.getAttribute("data-version")];for(const o of n){const t=String(o||"").trim();if(t==="")continue;const e=t.match(/^(\d+)/);if(!e)continue;const i=parseInt(e[1],10);if(Number.isFinite(i)&&i>0)return i}return null},C=n=>typeof window.t=="function"?window.t(w,n):n;export{w as A,d as M,r as V,O as a,u as b,a as c,c as d,p as n,s as o,C as t};
+//# sourceMappingURL=oc-compat-DyRpMLpS.chunk.mjs.map
